version 1.0.9 (#107)
* version 1.0.8

* ver 1.0.9 hotfix

+++ TarnishedTool/ViewModels/PlayerViewModel.cs
@@ -1118,6 +1118,7 @@ namespace TarnishedTool.ViewModels
         {
             // needed if player enables it while in Abyssal woods
             var playerIns = _playerService.GetPlayerIns();
+            if (playerIns == IntPtr.Zero) return;
             _spEffectService.RemoveSpEffect(playerIns, SpEffect.ForcedDismount);
 
             // needed for later area reloads
